requirements.txt
flask
flask-script
flask-sqlalchemy
flask-login
pyjwt
sqlalchemy_utils
